+++ src/main/groovy/com/nabilhachicha/nativedependencies/extension/NativeDependenciesExtension.groovy
@@ -20,11 +20,10 @@ import org.gradle.util.ConfigureUtil
 
 class NativeDependenciesExtension {
     final String CONFIGURATION_SEPARATOR = ":"
-    final def classifiers = ['armeabi', 'armeabi-v7a', 'x86', 'mips']
-
+    final def classifiers = ['armeabi', 'armeabi-v7a', 'arm64-v8a', 'x86', 'x86_64', 'mips', 'mips64']
     def dependencies = []
     /**
-     * set by a closure to let the user choose if he want to disable
+     * set by a closure to let the user choose if he/she wants to disable
      * prefixing the artifact with 'lib'
      */
     boolean addLibPrefixToArtifact = true