+++ src/test/groovy/com/nabilhachicha/nativedependencies/utils/ConfPath.groovy
@@ -1,0 +1,9 @@
+package com.nabilhachicha.nativedependencies.utils;
+
+/**
+ * Created by Nabil on 12/05/15.
+ */
+public class ConfPath {
+    String path
+    String jniLibsLocation
+}