+++ CHANGELOG.md
@@ -1,4 +1,13 @@
+Version 0.1.2 *(2015-05-12)*
+----------------------------
+- Add support for 64bit ABI
+
+
+Version 0.1.1 *(2014-08-17)*
+----------------------------
+- Add 'addLibPrefixToArtifact' to enable/disable the automatic prefix add to the native lib
+
 Version 0.1.0 *(2014-05-25)*
 ----------------------------
 
-Initial version for the 0.1.x Android native dependencies plugin.
+- Initial version for the 0.1.x Android native dependencies plugin.